+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `defaultDnt` to support setting the dnt flag for SLAS. Upgrade `commerce-sdk-isomorphic` to v3.1.1 [#1979](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1979)
 - Update logout helper to work for guest users [#1997](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1997)
+- Update `useCustomMutation` hook to accept body as mutate parameter [#2008](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2008)
 
 ## v3.0.1 (Sep 04, 2024)
 - Fixed an issue where the `expires` attribute in cookies, ensuring it uses seconds instead of days. [#1994](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1994)